split out dl stuff into its own file
